# shared-api/docs/openapi.yaml
openapi: 3.1.0
info:
  title: 'Melanie Data API '
  description: 'A high performance & centrally cached API service for premium bots. '
  version: ' commit e2ffc88 ok @ Sep 10 11:37 pm'
servers:
- url: https://dev.melaniebot.net
  description: montreal
paths:
...
  /api/web/telegram/{username}:
    get:
      tags:
      - web
      summary: Get Telegram User
      operationId: getTelegramUSer
      parameters:
      - required: true
        schema:
          type: string
          title: Username
        name: username
        in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TelegramProfileResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/web/screenshot:
    get:
      tags:
      - web
      summary: Screenshot Web
      description: Screenshot a webpage with Chrome
      operationId: screenshotweb
      parameters:
      - description: URL to screnshot
        required: true
        schema:
          type: string
          title: Url
          description: URL to screnshot
        name: url
        in: query
      - description: User ID requesting the screenshot.
        required: false
        schema:
          type: string
          title: User Id
          description: User ID requesting the screenshot.
        name: user_id
        in: query
      - description: Check the URL's domain against porn malware lists.
        required: false
        schema:
          type: boolean
          title: Safe
          description: Check the URL's domain against porn malware lists.
          default: true
        name: safe
        in: query
      - description: puppeteer wait method.
        required: false
        schema:
          allOf:
          - $ref: '#/components/schemas/PuppeteerLoadMethod'
          description: puppeteer wait method.
          default: domcontentloaded
        name: until
        in: query
      - description: 'Run the image result through AI content moderation  '
        required: false
        schema:
          type: boolean
          title: Nsfw Check
          description: 'Run the image result through AI content moderation  '
          default: true
        name: nsfw_check
        in: query
      - description: Waits for page to load, scrolls to bottom
        required: false
        schema:
          type: boolean
          title: Full Page
          description: Waits for page to load, scrolls to bottom
          default: true
        name: full_page
        in: query
      - description: png or jpeg
        required: false
        schema:
          type: string
          title: Image Type
          description: png or jpeg
          default: png
        name: image_type
        in: query
      - description: OK to return an image on failure or send HTTP error code
        required: false
        schema:
          type: boolean
          title: Imgerr
          description: OK to return an image on failure or send HTTP error code
          default: true
        name: imgerr
        in: query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/youtube/search:
    post:
      tags:
      - youtube
      summary: Search Youtube
      description: Search Youtube for videos given a query
      operationId: searchYouTube
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/YTSearchRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/YoutubeSearchResults'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/web/ip/{address}:
    get:
      tags:
      - web
      summary: Ip Lookup
      description: Get information on an IP address
      operationId: IPlookup
      parameters:
      - description: Full IP address or range to query info on
        required: true
        schema:
          type: string
          title: Address
          description: Full IP address or range to query info on
        name: address
        in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IPLookupResultResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/cashapp/{username}:
    get:
      tags:
      - web
      summary: Get Cashapp Profile
      description: Fetch a Cashapp user!
      operationId: getCashappprofile
      parameters:
      - required: true
        schema:
          type: string
          title: Username
        name: username
        in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CashappProfileResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/spotify/playlist/{playlist_id}:
    get:
      tags:
      - web
      summary: Load A Spotify Playlist
      operationId: loadaSpotifyplaylist
      parameters:
      - required: true
        schema:
          type: string
          title: Playlist Id
        name: playlist_id
        in: path
      - required: false
        schema:
          type: boolean
          title: Plain
          default: false
        name: plain
        in: query
      - required: false
        schema:
          type: boolean
          title: Transform
          default: false
        name: transform
        in: query
      - required: false
        schema:
          type: string
          title: Name
        name: name
        in: query
      - required: false
        schema:
          type: boolean
          title: Clean
          default: false
        name: clean
        in: query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Avatar:
      properties:
        image_url:
          type: string
          maxLength: 65536
          minLength: 1
          format: uri
          title: Image Url
        initial:
          type: string
          title: Initial
        accent_color:
          type: string
          title: Accent Color
      type: object
      title: Avatar
...
    CashAppProfile:
      properties:
        display_name:
          type: string
          title: Display Name
        formatted_cashtag:
          type: string
          title: Formatted Cashtag
        is_verified_account:
          type: boolean
          title: Is Verified Account
        rate_plan:
          type: string
          title: Rate Plan
        payment_button_type:
          type: string
          title: Payment Button Type
        country_code:
          type: string
          title: Country Code
        avatar:
          $ref: '#/components/schemas/Avatar'
      type: object
      title: CashAppProfile
    CashappProfileResponse:
      properties:
        profile:
          $ref: '#/components/schemas/CashAppProfile'
        qr_image_url:
          type: string
          title: Qr Image Url
      type: object
      title: CashappProfileResponse
...
    CountryCurrency:
      properties:
        code:
          type: string
          title: Code
        symbol:
          type: string
          title: Symbol
      type: object
      title: CountryCurrency
    CountryFlag:
      properties:
        emoji:
          type: string
          title: Emoji
        unicode:
          type: string
          title: Unicode
      type: object
      title: CountryFlag
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    IPLookupResultResponse:
      properties:
        ip:
          type: string
          title: Ip
        fraud_score:
          $ref: '#/components/schemas/IpScamScore'
        hostname:
          type: string
          title: Hostname
        city:
          type: string
          title: City
        region:
          type: string
          title: Region
        country:
          type: string
          title: Country
        country_name:
          type: string
          title: Country Name
        country_flag:
          $ref: '#/components/schemas/CountryFlag'
        country_currency:
          $ref: '#/components/schemas/CountryCurrency'
        loc:
          type: string
          title: Loc
        org:
          type: string
          title: Org
        postal:
          type: string
          title: Postal
        timezone:
          type: string
          title: Timezone
      type: object
      title: IPLookupResultResponse
...
    IpScamScore:
      properties:
        ip:
          type: string
          title: Ip
        score:
          type: string
          title: Score
        risk:
          type: string
          title: Risk
      type: object
      required:
      - ip
      - score
      - risk
      title: IpScamScore
...
    PuppeteerLoadMethod:
      type: string
      enum:
      - commit
      - load
      - networkidle
      - domcontentloaded
      title: PuppeteerLoadMethod
      description: An enumeration.
...
    TelegramProfileResponse:
      properties:
        username:
          type: string
          title: Username
        name:
          type: string
          title: Name
        description:
          type: string
          title: Description
        url:
          type: string
          title: Url
        avatar_url:
          type: string
          title: Avatar Url
      type: object
      required:
      - username
      title: TelegramProfileResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    YTSearchRequest:
      properties:
        query:
          type: string
          title: Query
      type: object
      required:
      - query
      title: YTSearchRequest
    YoutubeSearchResults:
      properties:
        results:
          items:
            $ref: '#/components/schemas/SearchResult'
          type: array
          title: Results
      type: object
      required:
      - results
      title: YoutubeSearchResults
...